- Reorganized Settings dialog. - Settings pages are now created on demand, which boosts the loading time of Settings dialog. - Settings are now saved only if the relevant page is created. - Added the option to display English titles instead of main titles in list. - Removed anime folders page from Settings dialog. - Removed ask for confirmation on exit option. - Removed save queued events on exit option. - Removed remember window position and size on exit option (only from Settings dialog).
git-svn-id: http://taiga.googlecode.com/svn/trunk@198 77add795-4bdf-f0f6-a574-f4d068491005

+++ dlg/dlg_about.cpp
@@ -160,7 +160,7 @@ void AboutDialog::OnPaint(HDC hdc, LPPAINTSTRUCT lpps) {
 
   // Paint background
   GetClientRect(&rect);
-  rect.left = ScaleX(48 * 1.5f);
+  rect.left = ScaleX(static_cast<int>(48 * 1.5f));
   dc.FillRect(rect, ::GetSysColor(COLOR_WINDOW));
 }
